Integrate chat engine and add tests
