+++ me.py
@@ -1,7 +1,8 @@
 import asyncio
 import random
+import re
 from typing import List, Tuple, Set, Optional
-from memory import tokenize, get_vocab, metrics
+from memory import tokenize, get_vocab, metrics, VerbGraph
 from rag import retrieve
 from method import train
 
@@ -95,8 +96,28 @@ class Engine:
         return reply_text
 
 
+class ChatEngine:
+    """Dialogue engine with verb-punctuation tracking."""
+
+    def __init__(self) -> None:
+        self.core = Engine()
+        self.graph = VerbGraph()
+
+    async def reply(self, message: str) -> str:
+        """Generate a reply and update verb graph."""
+        self.graph.add_sentence(message)
+        text = await self.core.reply(message)
+        match = re.search(r"(\b\w+\b)[.!?]$", text.strip())
+        if match:
+            verb = match.group(1)
+            punct = self.graph.preferred_punct(verb)
+            text = re.sub(r"[.!?]$", punct, text.strip())
+        self.graph.add_sentence(text)
+        return text
+
+
 if __name__ == '__main__':
-    bot = Engine()
+    bot = ChatEngine()
     try:
         while True:
             msg = input('> ')
